+++ plugins/SitesManager/API.php
@@ -14,6 +14,7 @@ use Piwik\Common;
 use Piwik\Container\StaticContainer;
 use Piwik\Date;
 use Piwik\Db;
+use Piwik\Log;
 use Piwik\Metrics\Formatter;
 use Piwik\Network\IPUtils;
 use Piwik\Option;
@@ -549,6 +550,10 @@ class API extends \Piwik\Plugin\API
         }
         $this->checkValidCurrency($currency);
 
+        // Revert the auto-escaping
+        // TODO remove when #4231 is fixed
+        $excludedUserAgents = Common::unsanitizeInputValue($excludedUserAgents);
+
         $url  = $urls[0];
         $urls = array_slice($urls, 1);
 
@@ -556,8 +561,8 @@ class API extends \Piwik\Plugin\API
                       'main_url' => $url);
 
         $bind['excluded_ips'] = $this->checkAndReturnExcludedIps($excludedIps);
-        $bind['excluded_parameters']  = $this->checkAndReturnCommaSeparatedStringList($excludedQueryParameters);
-        $bind['excluded_user_agents'] = $this->checkAndReturnCommaSeparatedStringList($excludedUserAgents);
+        $bind['excluded_parameters']  = $this->ensureLineReturnsSeparatedList($excludedQueryParameters);
+        $bind['excluded_user_agents'] = $this->ensureLineReturnsSeparatedList($excludedUserAgents);
         $bind['keep_url_fragment']    = $keepURLFragments;
         $bind['timezone']   = $timezone;
         $bind['currency']   = $currency;
@@ -728,12 +733,12 @@ class API extends \Piwik\Plugin\API
     }
 
     /**
-     * Checks that the submitted IPs (comma separated list) are valid
+     * Checks that the submitted IPs are valid
      * Returns the cleaned up IPs
      *
-     * @param string $excludedIps Comma separated list of IP addresses
+     * @param string $excludedIps Line return separated list of IP addresses
      * @throws Exception
-     * @return array of IPs
+     * @return string Line return separated list
      */
     private function checkAndReturnExcludedIps($excludedIps)
     {
@@ -741,7 +746,7 @@ class API extends \Piwik\Plugin\API
             return '';
         }
 
-        $ips = explode(',', $excludedIps);
+        $ips = explode("\n", $excludedIps);
         $ips = array_map('trim', $ips);
         $ips = array_filter($ips, 'strlen');
 
@@ -751,8 +756,7 @@ class API extends \Piwik\Plugin\API
             }
         }
 
-        $ips = implode(',', $ips);
-        return $ips;
+        return implode("\n", $ips);
     }
 
     /**
@@ -822,7 +826,7 @@ class API extends \Piwik\Plugin\API
      * Sets IPs to be excluded from all websites. IPs can contain wildcards.
      * Will also apply to websites created in the future.
      *
-     * @param string $excludedIps Comma separated list of IPs to exclude from being tracked (allows wildcards)
+     * @param string $excludedIps Line-return separated list of IPs to exclude from being tracked (allows wildcards)
      * @return bool
      */
     public function setGlobalExcludedIps($excludedIps)
@@ -879,12 +883,12 @@ class API extends \Piwik\Plugin\API
     /**
      * Returns the list of URL query parameters that are excluded from all websites
      *
-     * @return string Comma separated list of URL parameters
+     * @return string[]
      */
     public function getExcludedQueryParametersGlobal()
     {
         Piwik::checkUserHasSomeViewAccess();
-        return Option::get(self::OPTION_EXCLUDED_QUERY_PARAMETERS_GLOBAL);
+        return $this->lineReturnsSeparatedListToArray(Option::get(self::OPTION_EXCLUDED_QUERY_PARAMETERS_GLOBAL));
     }
 
     /**
@@ -892,27 +896,30 @@ class API extends \Piwik\Plugin\API
      * all websites. If a visitor's user agent string contains one of these substrings,
      * their visits will not be included.
      *
-     * @return string Comma separated list of strings.
+     * @return string[]
      */
     public function getExcludedUserAgentsGlobal()
     {
         Piwik::checkUserHasSomeAdminAccess();
-        return Option::get(self::OPTION_EXCLUDED_USER_AGENTS_GLOBAL);
+        return $this->lineReturnsSeparatedListToArray(Option::get(self::OPTION_EXCLUDED_USER_AGENTS_GLOBAL));
     }
 
     /**
      * Sets list of user agent substrings to look for when excluding visits. For more info,
      * @see getExcludedUserAgentsGlobal.
      *
-     * @param string $excludedUserAgents Comma separated list of strings. Each element is trimmed,
-     *                                   and empty strings are removed.
+     * @param string $excludedUserAgents Line-return separated list. Each element is trimmed,
+     *                                   enclosing quotes are removed and empty strings too.
      */
     public function setGlobalExcludedUserAgents($excludedUserAgents)
     {
         Piwik::checkUserHasSuperUserAccess();
 
-        // update option
-        $excludedUserAgents = $this->checkAndReturnCommaSeparatedStringList($excludedUserAgents);
+        // Revert the auto-escaping
+        // TODO remove when #4231 is fixed
+        $excludedUserAgents = Common::unsanitizeInputValue($excludedUserAgents);
+
+        $excludedUserAgents = $this->ensureLineReturnsSeparatedList($excludedUserAgents);
         Option::set(self::OPTION_EXCLUDED_USER_AGENTS_GLOBAL, $excludedUserAgents);
 
         // make sure tracker cache will reflect change
@@ -983,13 +990,13 @@ class API extends \Piwik\Plugin\API
      * Sets list of URL query parameters to be excluded on all websites.
      * Will also apply to websites created in the future.
      *
-     * @param string $excludedQueryParameters Comma separated list of URL query parameters to exclude from URLs
+     * @param string $excludedQueryParameters Line-returns separated list of URL query parameters to exclude from URLs
      * @return bool
      */
     public function setGlobalExcludedQueryParameters($excludedQueryParameters)
     {
         Piwik::checkUserHasSuperUserAccess();
-        $excludedQueryParameters = $this->checkAndReturnCommaSeparatedStringList($excludedQueryParameters);
+        $excludedQueryParameters = $this->ensureLineReturnsSeparatedList($excludedQueryParameters);
         Option::set(self::OPTION_EXCLUDED_QUERY_PARAMETERS_GLOBAL, $excludedQueryParameters);
         Cache::deleteTrackerCache();
         return true;
@@ -998,12 +1005,12 @@ class API extends \Piwik\Plugin\API
     /**
      * Returns the list of IPs that are excluded from all websites
      *
-     * @return string Comma separated list of IPs
+     * @return string[]
      */
     public function getExcludedIpsGlobal()
     {
         Piwik::checkUserHasSomeAdminAccess();
-        return Option::get(self::OPTION_EXCLUDED_IPS_GLOBAL);
+        return $this->lineReturnsSeparatedListToArray(Option::get(self::OPTION_EXCLUDED_IPS_GLOBAL));
     }
 
     /**
@@ -1119,6 +1126,10 @@ class API extends \Piwik\Plugin\API
             throw new Exception("website id = $idSite not found");
         }
 
+        // Revert the auto-escaping
+        // TODO remove when #4231 is fixed
+        $excludedUserAgents = Common::unsanitizeInputValue($excludedUserAgents);
+
         // Build the SQL UPDATE based on specified updates to perform
         $bind = array();
 
@@ -1157,8 +1168,8 @@ class API extends \Piwik\Plugin\API
             $bind['ts_created'] = Date::factory($startDate)->getDatetime();
         }
         $bind['excluded_ips'] = $this->checkAndReturnExcludedIps($excludedIps);
-        $bind['excluded_parameters'] = $this->checkAndReturnCommaSeparatedStringList($excludedQueryParameters);
-        $bind['excluded_user_agents'] = $this->checkAndReturnCommaSeparatedStringList($excludedUserAgents);
+        $bind['excluded_parameters'] = $this->ensureLineReturnsSeparatedList($excludedQueryParameters);
+        $bind['excluded_user_agents'] = $this->ensureLineReturnsSeparatedList($excludedUserAgents);
 
         if (!is_null($keepURLFragments)) {
             $keepURLFragments = (int)$keepURLFragments;
@@ -1214,18 +1225,26 @@ class API extends \Piwik\Plugin\API
         $this->getModel()->updateSiteCreatedTime($idSites, $minDateSql);
     }
 
-    private function checkAndReturnCommaSeparatedStringList($parameters)
+    private function ensureLineReturnsSeparatedList($list)
     {
-        $parameters = trim($parameters);
-        if (empty($parameters)) {
+        $list = trim($list);
+        if (empty($list)) {
             return '';
         }
 
-        $parameters = explode(',', $parameters);
-        $parameters = array_map('trim', $parameters);
-        $parameters = array_filter($parameters, 'strlen');
-        $parameters = array_unique($parameters);
-        return implode(',', $parameters);
+        $list = explode("\n", $list);
+        // Remove enclosing quotes
+        $list = array_map(function ($item) {
+            return trim(trim($item), '"');
+        }, $list);
+        $list = array_filter($list, 'strlen');
+        $list = array_unique($list);
+        return implode("\n", $list);
+    }
+
+    private function lineReturnsSeparatedListToArray($list)
+    {
+        return array_filter(explode("\n", $list), 'strlen');
     }
 
     /**